Feat : 랭킹 버튼 추가 및 인벤토리 버그 수정

+++ Assets/Scripts/Town/UI Scripts/UIPlayer.cs
@@ -150,7 +150,7 @@ public class UIPlayer : MonoBehaviour
             DeActiveAP();
     }
 
-    public void SetMoveSpeed(int moveSpeed, bool hasAP)
+    public void SetMoveSpeed(float moveSpeed, bool hasAP)
     {
         moveSpeedText.text = moveSpeed.ToString();
         if (!hasAP)
